refactor: extract the logic to load/store users
