+++ dinstaller-lib/src/store.rs
@@ -1,14 +1,16 @@
-use crate::install_settings::{InstallSettings, SoftwareSettings, StorageSettings, UserSettings};
+mod users;
+
+use crate::install_settings::{InstallSettings, SoftwareSettings, StorageSettings};
 use crate::software::SoftwareClient;
 use crate::storage::StorageClient;
-use crate::users::{FirstUser, UsersClient};
+use crate::store::users::UsersStore;
 use std::{default::Default, error::Error};
 
 /// Loading and storing the settings in the D-Bus service
 ///
 /// This struct uses the default connection built by [connection function](super::connection).
 pub struct Store<'a> {
-    users_client: UsersClient<'a>,
+    users: UsersStore<'a>,
     software_client: SoftwareClient<'a>,
     storage_client: StorageClient<'a>,
 }
@@ -16,7 +18,7 @@ pub struct Store<'a> {
 impl<'a> Store<'a> {
     pub fn new() -> Result<Self, zbus::Error> {
         Ok(Self {
-            users_client: UsersClient::new(super::connection()?)?,
+            users: UsersStore::new(super::connection()?)?,
             software_client: SoftwareClient::new(super::connection()?)?,
             storage_client: StorageClient::new(super::connection()?)?,
         })
@@ -24,7 +26,6 @@ impl<'a> Store<'a> {
 
     /// Loads the installation settings from the D-Bus service
     pub fn load(&self) -> Result<InstallSettings, Box<dyn Error>> {
-        let first_user = self.users_client.first_user()?;
         let product = self.software_client.product()?;
 
         let settings = InstallSettings {
@@ -32,12 +33,7 @@ impl<'a> Store<'a> {
             software: SoftwareSettings {
                 product: Some(product),
             },
-            user: UserSettings {
-                user_name: Some(first_user.user_name),
-                autologin: Some(first_user.autologin),
-                full_name: Some(first_user.full_name),
-                password: Some(first_user.password),
-            },
+            user: self.users.load()?,
         };
         Ok(settings)
     }
@@ -45,21 +41,8 @@ impl<'a> Store<'a> {
     /// Stores the given installation settings in the D-Bus service
     pub fn store(&self, settings: &InstallSettings) -> Result<(), Box<dyn Error>> {
         self.store_software_settings(&settings.software)?;
-        self.store_user_settings(&settings.user)?;
+        self.users.store(&settings.user)?;
         self.store_storage_settings(&settings.storage)?;
-        Ok(())
-    }
-
-    fn store_user_settings(&self, settings: &UserSettings) -> Result<(), Box<dyn Error>> {
-        // fixme: improve
-        let first_user = FirstUser {
-            user_name: settings.user_name.clone().unwrap_or_default(),
-            full_name: settings.full_name.clone().unwrap_or_default(),
-            autologin: settings.autologin.unwrap_or_default(),
-            password: settings.password.clone().unwrap_or_default(),
-            ..Default::default()
-        };
-        self.users_client.set_first_user(&first_user)?;
         Ok(())
     }
